+++ src/components/popup/popup.ts
@@ -1,4 +1,7 @@
+<<<<<<< HEAD
 import SlPopup from './popup.component.js';
 export * from './popup.component.js';
 export default SlPopup;
 SlPopup.define('sl-popup');
+=======
+>>>>>>> next